backend/requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
psycopg2-binary==2.9.10
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3